Extract moderation events and session effects
