+++ late-ssh/src/moderation/mod.rs
@@ -1,3 +1,5 @@
 pub mod command;
+pub mod event;
 pub mod policy;
 pub mod service;
+pub mod session_effects;
